Move `ExpressionSerializer` from `opensearch` to `core`.
